+++ matrix.go
@@ -3,17 +3,17 @@ package main
 type (
 	Matrix struct {
 		Data []float32
-		Rows int
-		Cols int
+		Rows uint32
+		Cols uint32
 	}
 )
 
-func SingletonMatrix(rows int, data []float32) *Matrix {
+func SingletonMatrix(rows uint32, data []float32) *Matrix {
 	return NewMatrix(rows, 1, data)
 }
 
-func NewMatrix(rows, cols int, data []float32) *Matrix {
-	if len(data) != rows*cols {
+func NewMatrix(rows, cols uint32, data []float32) *Matrix {
+	if len(data) != int(rows*cols) {
 		panic("Wrong matrix dimensions")
 	}
 
@@ -24,27 +24,27 @@ func NewMatrix(rows, cols int, data []float32) *Matrix {
 	}
 }
 
-func (m *Matrix) Size() int {
+func (m *Matrix) Size() uint32 {
 	return m.Rows * m.Cols
 }
 
-func (m *Matrix) Get(row, col int) float32 {
+func (m *Matrix) Get(row, col uint32) float32 {
 	return m.Data[col*m.Rows+row]
 }
 
-func (m *Matrix) Set(row, col int, v float32) {
+func (m *Matrix) Set(row, col uint32, v float32) {
 	m.Data[col*m.Rows+row] = v
 }
 
-func (m *Matrix) AddTo(row, col int, v float32) {
+func (m *Matrix) AddTo(row, col uint32, v float32) {
 	m.Data[col*m.Rows+row] += v
 }
 
 func (m *Matrix) Product(fst, snd *Matrix) {
-	for i := 0; i < fst.Rows; i++ {
-		for j := 0; j < snd.Cols; j++ {
+	for i := uint32(0); i < fst.Rows; i++ {
+		for j := uint32(0); j < snd.Cols; j++ {
 			m.Set(j, i, 0)
-			for k := 0; k < fst.Cols; k++ {
+			for k := uint32(0); k < fst.Cols; k++ {
 				m.AddTo(j, i, fst.Get(i, k)*snd.Get(k, j))
 			}
 		}
@@ -52,16 +52,16 @@ func (m *Matrix) Product(fst, snd *Matrix) {
 }
 
 func (m *Matrix) Add(other *Matrix, fn func(float32) float32) {
-	for i := 0; i < m.Rows; i++ {
-		for j := 0; j < m.Cols; j++ {
+	for i := uint32(0); i < m.Rows; i++ {
+		for j := uint32(0); j < m.Cols; j++ {
 			m.Set(j, i, fn(m.Get(i, j)+other.Get(i, j)))
 		}
 	}
 }
 
 func (m *Matrix) Subtract(other *Matrix) {
-	for i := 0; i < m.Rows; i++ {
-		for j := 0; j < m.Cols; j++ {
+	for i := uint32(0); i < m.Rows; i++ {
+		for j := uint32(0); j < m.Cols; j++ {
 			m.Set(j, i, m.Get(i, j)-other.Get(i, j))
 		}
 	}
